relocated api-service to shared
