+++ src/models/index.js
@@ -1,4 +1,3 @@
-export { ApiService } from './api-service/index.js';
 export {
   mentorsReducer, loadMentors, loadMentorsWatcher, mentorsIsLoadingSelector, mentorsDataSelector,
 } from './mentors/index.js';
